updated test 2282 and removed 2341
