+++ apps/src/tests/index.ts
@@ -112,7 +112,6 @@ export { default as Test2271 } from './Test2271';
 export { default as Test2282 } from './Test2282';
 export { default as Test2317 } from './Test2317';
 export { default as Test2332 } from './Test2332';
-export { default as Test2341 } from './Test23341';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 export { default as TestHeader } from './TestHeader';
 export { default as TestModalNavigation } from './TestModalNavigation';
